+++ lib/prism/merge/begin_node_clause_body_merger.rb
@@ -40,6 +40,7 @@ module Prism
           signature_generator: merger.instance_variable_get(:@raw_signature_generator),
           preference: merger.preference,
           add_template_only_nodes: merger.add_template_only_nodes,
+          remove_template_missing_nodes: merger.remove_template_missing_nodes,
           freeze_token: merger.freeze_token,
           max_recursion_depth: merger.max_recursion_depth,
           current_depth: merger.instance_variable_get(:@current_depth) + 1,